Pongo casos de Trenes dentro del Package Trenes
